chore: divide calendar component
