+++ src/components/calendar/dateArea.tsx
@@ -1,0 +1,94 @@
+/** @jsx jsx */
+import React from "react";
+import styled from "@emotion/styled";
+import { jsx } from "@emotion/core";
+import { Dayjs } from "dayjs";
+import WeekArea from "./weekArea";
+import { range } from "../../common/util";
+
+interface Props {
+  date: Dayjs;
+  setDate: React.Dispatch<React.SetStateAction<Dayjs>>;
+}
+
+const DateArea = styled("div")`
+  display: flex;
+  flex-direction: column;
+`;
+
+const DayOfWeekArea = styled("div")`
+  display: flex;
+  justify-content: space-around;
+  width: 100%;
+  padding: 10px 0;
+`;
+
+interface DayOfWeekTitleProps {
+  sunday?: boolean;
+}
+
+const DayOfWeekTitle = styled("span")<DayOfWeekTitleProps>(({ sunday }: DayOfWeekTitleProps) => ({
+  display: "flex",
+  fontSize: "14px",
+  width: "36px",
+  justifyContent: "center",
+  color: sunday ? "rgb(208,2,27)" : "rgb(44,44,44)"
+}));
+
+const DayArea = styled("div")`
+  display: flex;
+  flex-direction: column;
+`;
+
+export default function CalendarDateArea(props: Props) {
+  const dayOfWeek = ["sun", "mon", "tue", "wed", "thu", "fri", "sat"];
+  const { date, setDate } = props;
+
+  const getDateArr = () => {
+    const numberOfDay = date.endOf("month").diff(date.startOf("month"), "day") + 1;
+    const numberOfprevMonthDay = date
+      .startOf("month")
+      .diff(date.startOf("month").startOf("week"), "day");
+    const numberOfNextMonthDay = date
+      .endOf("month")
+      .endOf("week")
+      .diff(date.endOf("month"), "day");
+
+    let result: (null | Dayjs)[] = range(0, numberOfprevMonthDay).map(() => null);
+
+    result = range(0, numberOfDay).reduce((acc, _, idx) => {
+      return [...acc, date.startOf("month").add(idx, "day")];
+    }, result);
+
+    return [...result, ...range(0, numberOfNextMonthDay).map(() => null)];
+  };
+
+  const cut = (data: (null | Dayjs)[], numb: number) => {
+    const len = Math.ceil(data.length / numb);
+    const res = [];
+    for (let i = 0; i < len; i += 1) {
+      res.push(data.slice(i * numb, i * numb + numb));
+    }
+
+    return res;
+  };
+
+  const dateArr = cut(getDateArr(), 7);
+
+  return (
+    <DateArea>
+      <DayOfWeekArea>
+        {dayOfWeek.map((day, dayIndex) => (
+          <DayOfWeekTitle key={`daytitle_${day}`} sunday={dayIndex === 0}>
+            {day}
+          </DayOfWeekTitle>
+        ))}
+      </DayOfWeekArea>
+      <DayArea>
+        {dateArr.map((week, index) => (
+          <WeekArea week={week} key={`weekArea_${index}`} date={date} setDate={setDate} />
+        ))}
+      </DayArea>
+    </DateArea>
+  );
+}
